change long to BigInteger

+++ src/main/java/Abstract/Herbivorous.java
@@ -5,6 +5,7 @@ import GameObjects.GameField;
 import GameObjects.Types;
 import Interfaces.GameObject;
 
+import java.math.BigInteger;
 import java.util.concurrent.ArrayBlockingQueue;
 
 
@@ -23,7 +24,7 @@ public abstract class Herbivorous extends Animal{
                        int food = this.getMaxFood();   // set weight food = this maxFood
                        O.setWeight(O.getWeight() - food); // set Grass weight = Grass weight - food
                        synchronized (gameField) {
-                           gameField.setEatenGrassWeight(food); // set Map for Grass
+                           gameField.setEatenGrassWeight(BigInteger.valueOf(food)); // set Map for Grass
                        }
                        this.setWeight(this.getWeight() + food); // set this weight = this weight + food
                          if(this.getWeight() > this.getMaxWeight()){ // check if this weight > this max weight
@@ -40,7 +41,7 @@ public abstract class Herbivorous extends Animal{
                        if(this.getWeight() + food < this.getMaxWeight()) { // check if this may eat all Grass
                            O.setWeight(0); // set Grass weight = 0
                            synchronized (gameField) {
-                               gameField.setEatenGrassWeight(food); // set Map Grass
+                               gameField.setEatenGrassWeight(BigInteger.valueOf(food)); // set Map Grass
                            }
                            this.setWeight(this.getWeight() + food); // set this weight = this weight + food
                            System.out.println(this + " ate " + O); // print eat
@@ -50,7 +51,7 @@ public abstract class Herbivorous extends Animal{
                                 food = this.getMaxWeight() - this.getWeight(); // set food = this maxWeight - this weight
                                O.setWeight(O.getWeight() - food);//set Grass weight = Grass weight - food
                                synchronized (gameField) {
-                                   gameField.setEatenGrassWeight(food); // set Map Grass
+                                   gameField.setEatenGrassWeight(BigInteger.valueOf(food)); // set Map Grass
                                }
                                this.setWeight(this.getMaxWeight()); // set this weight = this maxWeight
                                System.out.println(this + " ate " + O); // print eat
@@ -59,7 +60,7 @@ public abstract class Herbivorous extends Animal{
                                food = O.getWeight(); // food  = Grass weight
                                O.setWeight(0); // set Grass weight = 0
                                synchronized (gameField) {
-                                   gameField.setEatenGrassWeight(food); // set Map Grass
+                                   gameField.setEatenGrassWeight(BigInteger.valueOf(food)); // set Map Grass
                                }
                                this.setWeight(this.getWeight() + food); // set this weight  = this weight + food
                                System.out.println(this + " ate " + O); // print eat
