+++ src/main/java/GameObjects/Cell.java
@@ -3,11 +3,12 @@ package GameObjects;
 import Interfaces.GameObject;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.util.concurrent.ArrayBlockingQueue;
 
 public class Cell implements Serializable {
     private  int capacity;
-    private long id;
+    private BigInteger id;
     private  ArrayBlockingQueue<GameObject> queue;
 
     public Cell() {
@@ -16,7 +17,7 @@ public class Cell implements Serializable {
         queue = new ArrayBlockingQueue<>(capacity);
     }
 
-    public Cell(long id) {
+    public Cell(BigInteger id) {
         this.id = id;
         queue = new ArrayBlockingQueue<>(capacity);
     }
@@ -25,11 +26,11 @@ public class Cell implements Serializable {
         return queue;
     }
 
-    public long getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
